+++ .gitignore
@@ -10,7 +10,7 @@
 !/.idea/codeStyles/codeStyleConfig.xml
 !/.idea/dictionaries
 /.idea/dictionaries/*
-!/.idea/dictionaries/tsundokuDictionary.xml
+!/.idea/dictionaries/tsundokunDictionary.xml
 !/.idea/icon.png
 .DS_Store
 /build
