--- .idea/dictionaries/tsundokuDictionary.xml
@@ -1,7 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="tsundokuDictionary">
-    <words>
-      <w>tsundoku</w>
-    </words>
-  </dictionary>
-</component>
