+++ app/src/androidTest/java/com/example/tsundokun/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.tsundoku
+package com.example.tsundokun
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -17,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.tsundoku", appContext.packageName)
+        assertEquals("com.example.tsundokun", appContext.packageName)
     }
 }
