+++ app/src/main/java/com/example/tsundokun/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.tsundoku
+package com.example.tsundokun
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -7,15 +7,16 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import com.example.tsundoku.ui.NavGraphs
-import com.example.tsundoku.ui.theme.TsundokuTheme
+import com.example.tsundokun.ui.NavGraphs
+//import com.example.tsundokun.ui.NavGraphs
+import com.example.tsundokun.ui.theme.TsundokunTheme
 import com.ramcosta.composedestinations.DestinationsNavHost
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            TsundokuTheme {
+            TsundokunTheme {
                 // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
