+++ app/src/main/java/com/example/tsundokun/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.example.tsundoku.ui.theme
+package com.example.tsundokun.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -38,7 +38,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun TsundokuTheme(
+fun TsundokunTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
