+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">tsundoku</string>
+    <string name="app_name">tsundokun</string>
 </resources>
