+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "tsundoku"
+rootProject.name = "tsundokun"
 include ':app'
